Moved events to core module
