+++ hero-quest-core/src/main/java/br/com/miguelmf/heroquest/core/events/BattleTurnComputedEvent.java
@@ -1,4 +1,4 @@
-package br.com.miguelmf.heroquest.events;
+package br.com.miguelmf.heroquest.core.events;
 
 public class BattleTurnComputedEvent {
